カテゴリーセルを別ファイルにする
